pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "euroc_converter"
version = "0.0.1"
description = "Script to convert EuRoC dataset to TUM or Replica format using dept estimation"
readme = "README.md"
authors = [{ name = "texnomann", email = "texnoman@itmo.ru" }]
dependencies = [
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "numpydantic",
    "scipy",
    "pyyaml",
    "click",
    "gdown",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "pypose",
    "pandas",
]

license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


[project.scripts]
euroc-converter = "euroc_converter.run:main"

[tool.setuptools]
package-dir = { "" = "src" }

[project.optional-dependencies]
dev = ["pytest>=7.4", "black>=23.11", "ruff>=0.1.6"]


[tool.black]
line-length = 79
include = '\.pyi?$'
